# server/openapi.yaml
openapi: 3.0.3
info:
  title: Slinger Cloud API
  version: 2.0.0
  description: |-
    Generated from the zod schemas of the routes registered in `server/src`; do not edit by hand (`npm run openapi`).

    Auth: `Authorization: Bearer <access token>` (desktop) or the `slinger_session` httpOnly cookie (dashboard). Cookie-authenticated unsafe requests must send `X-CSRF-Token` (from `POST /v1/auth/browser/login` or `GET /v1/auth/browser/session`).

    Every response carries `X-Request-Id`, echoed in `error.request_id`. List endpoints use cursor pagination (`?cursor=&limit=&order=`) and return `{ items, page: { next_cursor, has_more } }`.

    Behind the bundled Caddy proxy the same routes are reachable at `/v1/...` and, for the dashboard, `/api/v1/...`.
servers:
  - url: http://localhost:8080
    description: Direct to the server container
...
paths:
...
  /device:
    get:
      operationId: getDevice
      summary: Device login approval page (HTML). Open from `verification_uri_complete`.
      tags:
        - Auth
      security: []
      responses:
        "200":
          description: OK
          content:
            text/html:
              schema:
                type: string
      parameters:
        - name: user_code
          in: query
          required: false
          schema:
            type: string
            maxLength: 16
    post:
      operationId: postDevice
      summary: Submit code + email + password (urlencoded form) to approve a device login
      tags:
        - Auth
      security: []
      responses:
        "200":
          description: OK
          content:
            text/html:
              schema:
                type: string
        "400":
          description: Invalid request (zod validation failed, malformed JSON, bad cursor)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      description: Rate limited per IP+email pair like the login endpoints. Accounts that must change an admin-issued password are refused (403 page with instructions) until they have done so in the dashboard.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                user_code:
                  type: string
                  minLength: 4
                  maxLength: 16
                email:
                  type: string
                  format: email
                  maxLength: 254
                password:
                  type: string
                  minLength: 1
                  maxLength: 256
              required:
                - user_code
                - email
                - password
              additionalProperties: false
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: slinger_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with cookieAuth on POST/PUT/PATCH/DELETE.
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - unauthenticated
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - workspace_access_denied
                - version_mismatch
                - invite_invalid
                - join_request_not_allowed
                - sync_conflict
                - csrf_invalid
                - origin_not_allowed
                - password_change_required
                - internal_error
            message:
              type: string
            details:
              type: object
              additionalProperties: {}
            request_id:
              type: string
          required:
            - code
            - message
            - details
            - request_id
          additionalProperties: false
      required:
        - error
      additionalProperties: false